Add Glooper code as submodule.
Also got rid of the old revision of the Gloop I instructions and renamed
the current version appropriately. I think that ostensibly glooper
should go in another directory (as an instrument rather than a
composition), but this seems fine for now - at least until we use it for
a new composition.